==> code/preprocess/2.make_before_matching_list.ipynb <==
import pandas as pd

# number of all patients :272,337
#   DME & Health check-up data is available until the onset: 2,368
#   non DME   : 265,475

kanja_summary = pd.read_csv("../data/kanja_summary.csv")
kanja_summary.rename(columns={'加入者id':'pid',
                              '加入者性別':'sex',
                              '観察開始時年齢':'age',
                              '観察期間_月':'period'},inplace=True)
cols = ['pid','is_mac','sex','age','period','durling_DM','months_until_first_kenshin','kenshin_times']
kanja_summary['sex'] = kanja_summary['sex'].replace({'男性': 'M', '女性': 'F'})

before_matching_mac_list = kanja_summary.loc[~kanja_summary.kenshin_times_before_onset.isna(),cols]
before_matching_ctl_list = kanja_summary.loc[kanja_summary.possible_control==1,cols]

before_matching_list = pd.concat([before_matching_mac_list,before_matching_ctl_list],axis=0)
before_matching_list.fillna(0,inplace=True)
before_matching_list.to_csv('../data/before_matching_list.csv',index=False)
