==> code/preprocess/6.interpolate.ipynb <==
import pandas as pd

df = pd.read_csv('./data/before_interpolate.csv')

medians = df.iloc[:,1:].median()

df_copy = df.copy()
df_new = pd.DataFrame([])
for pid in df_copy.加入者id.unique():
    df_some_pid = df_copy.loc[df_copy.加入者id==pid,:]
    df_some_pid.iloc[:1,1:] = df_some_pid.iloc[:1,1:].fillna(medians)
    df_filled_ffill = df_some_pid.ffill()
    df_new = pd.concat([df_new,df_filled_ffill])

df_new.to_csv('./data/after_interpolate.csv',index=False)
